8-5-2015 improve ng.ckeditor
First of all, you don't need to change anything to use this version.
Upgrade made:
1. allow width and resizeEnabled be defined
2. add an notInitial lock to detect if function:  editor.setData has
been called or not.
3. If editor.setData has been called for first time, then inside scope,
the original code will keep calling editor.setData which will put
cursor to the beginning of the textarea.....no body can type
anything....
4. editor.setData will be called when it get focused which will
maintain the cursor in the right position
5. .min.js  should NOT shrink variables. Otherwise, it will cause error
said: Uncaught ReferenceError: scope is not defined