Fixes mouse position on unfocused click
When clicking on the CKeditor the first time, the position of the click
was lost due to a setData call which reset it.

The fix is to only set the data if it is different to the current content
of CKeditor. For most cases this avoids the position reset.

Note that this is more of a triage than a true fix.